Fix ownership of DhLinks in DhBase
Previously there were two collections of lists in DhBase: priv->keywords, and
priv->book_tree.  Neither of these lists had clear ownership of the DhLinks that
they held.  This patch adds a ref to the DhLink before adding it to each list,
and then unrefs the items from each list when destroying DhBase, in order to
prevent leaking of all of these objects.

https://bugzilla.gnome.org/show_bug.cgi?id=596808