Clean up application windows on 'Quit' action
Previously, when the 'Quit' action was activated, we just called
gtk_main_quit(), but that doesn't give the application a chance to clean up all
open windows, etc.  This resulted in a leak at exit (I know, not a big deal to
leak at exit, but still...) because each open window didn't get a chance to
unref their priv->base member, so the g_object_unref(base) in main() did not
actually result in the base object being destroyed, so the DhBase destructor was
never being called when the application was exited using the 'Quit' action (it
was behaving correctly when the windows were closed by the window manager).

https://bugzilla.gnome.org/show_bug.cgi?id=596808