Fix zoom in/out actions
Use double values in the zoom-levels array, using the values defines in
ephy-zoom control, and get the array index for the nearest zoom level
with the same approach implemented in ephy-zoom control too. This also
simplies the code, since we can use the zoom levels defined in
zoom-levels array directly as the WebKitWebView zoom level.

https://bugzilla.gnome.org/show_bug.cgi?id=658796