book-tree: avoid expanding selected items
When moving through the treeview with the keyboard we were expanding the items
as we moved. This was an artifact due to the webview notifying us back that we
had to re-select a given path, and while doing that we would expand_to_path()
which expands not only the parents but also the current item.

So, just avoid re-selection and expansion if we're already selected there.

Should fix https://bugzilla.gnome.org/show_bug.cgi?id=673320