book-manager: Don't pass separator to g_build_filename
g_build_filename automatically adds G_DIR_SEPARATOR_S and its parameters should
be the components of the path. Passing G_DIR_SEPARATOR_S to it will add another
separator to the start of the path.

Probably a confusion with g_build_path, which needs the separator as its first
parameter.