keyword-model: always look in other books for an exact match
We were performing a search in the current book first, giving priority to the
results of the current book, even if they were not prefix results. But, this
meant that if the exact match was in another book and the result list already
reached MAX_HITS, we wouldn't be showing the exact match.

This patch changes the behaviour of the search result building, so that we
always look for an exact match in other books as well as the current book, and
if the exact match is found in other books, we give priority to those.

https://bugzilla.gnome.org/show_bug.cgi?id=728916