Don't crash with empty .devhelp* files
Call g_markup_parse_context_end_parse after we are done reading the
file to catch empty files among other things. This is the only method
that raises G_MARKUP_ERROR_EMPTY, so it is particularly necessary for
this case.

It was necessary to remove parser_error_cb because we should not be
freeing the context inside it and dh_parser_free would have done it
for us anyway. The error callback is called from set_error_literal
inside g_markup_parse_context_end_parse, and invalidating the context
can lead to bad things.

Construction of a DhBook instance can fail due to parsing errors, so
be careful about that.

https://bugzilla.gnome.org/show_bug.cgi?id=742447