Use bitfields for booleans in structs
It saves a little bit of memory. It's better to do it each time, so when
another boolean is added it's directly beneficial (and we don't need to
think all the time: should I write a bitfield of "gboolean"? Is the
struct heavily used or not? etc. It's simpler to always apply the same
convention).