keyword-model: use GQueues to concatenate lists
It's a more appropriate data structure for our needs. With the util
function it's almost the same code to concatenate the lists.

And it's nice to get rid of the new_list parameter of
keyword_model_search_book() (the function is more self-contained that
way, it doesn't need the list of the caller).