link: sort pages before other link types
Before this commit, the g_list_sort() in keyword_model_search_books()
worked by chance. When keyword_model_search_books() has been modified to
use a GQueue, and that elements are pushed on the tail instead of being
prepended, it has changed the order of some equal symbols.

But we want page links before other link types, for symbols with the
same name. For example "GObject" can refer to a page, or a
typedef/struct.

https://bugzilla.gnome.org/show_bug.cgi?id=749797