sidebar: check that the hitlist is visible in entry_key_press_event_cb
There was a critical message when the search entry is empty and we press
Ctrl+tab:

Gtk-CRITICAL **: gtk_widget_event: assertion 'WIDGET_REALIZED_FOR_EVENT (widget, event)' failed

Which is normal because when the search entry is empty, the hitlist
isn't shown, the book tree is visible instead.

Similarly, when Return was pressed when the search entry is empty, it
searched the previous content of the hitlist.

https://bugzilla.gnome.org/show_bug.cgi?id=756341