sidebar: fix bug in set_search_string()
The following command didn't work:
$ devhelp --search "gtk_ translation"

It showed the book tree, not the hitlist. Because
sidebar_entry_changed_cb() hides or shows the hitlist and the book tree,
depending on whether the entry is empty or not.

It's amazing the number of bugs that we find just by reading the code.

https://bugzilla.gnome.org/show_bug.cgi?id=756341