sidebar: always send link-selected signal
Storing the selected_link is not needed. In every case where the
link-selected signal is sent, it is better to send it regardless of
whether the previous signal emission provided the same link. By
re-sending the signal with the same link, the page scrolls at the right
location (if the user has scrolled within the page).

https://bugzilla.gnome.org/show_bug.cgi?id=756341