book-tree: nicer behavior when changing the group-by-language setting
1. launch devhelp, select a page
2. go to the preferences and change the group-by-language setting
   -> the last book in the tree was opened and expanded, why?

When gtk_tree_store_clear() was called after changing the setting, the
GtkTreeView selection changed for each row deleted from the store, one
by one. So there was a signal to open each book, and the last one wins.

Now disconnect the model from the view when clearing the store, so the
signals are not emitted, so the WebkitWebView doesn't change. Ideally I
think it would be better to clear the WebkitWebView as well, to have a
blank page after changing the setting. But it would be a bit more
complicated, and the new behavior is already much better than before.

Also, book_tree_init_selection() must be called.