window: implement scroll zoom
This patch enables to change the zoom level by scrolling and holding the
ctrl modifier.

zoom_in_cb() and zoom_out_cb() used largely the same code for changing
the zoom level. It has been moved out to a separate function
(bump_zoom_level()), which takes the amount by which to "bump" the zoom
level.

https://bugzilla.gnome.org/show_bug.cgi?id=763942