BookManager: add get_singleton() and free_singleton()
The implementation is like in DhSettings.

It makes sense to make DhBookManager a singleton, because it's the
top-level class containing the data. And we want to load the data just
once, and access the same data everywhere. There is no need to create
several DhBookManager instances.

Currently a lot of classes take a DhBookManager reference, either with a
construct-only property or a set_book_manager() function. So it'll be
possible to simplify the code, by just calling
dh_book_manager_get_singleton() where we need it.