Window: first step to make it private
DhWindow is not re-usable. It depends on DhApp, and DhApp is a subclass
of GtkApplication. An application can have only one GtkApplication
instance, so DhApp in the current state doesn't make sense as a library.

This is the first step to make DhWindow private, it removes the class
from the public docs.

The gtk-doc comments are removed, because it is not really useful, and
in other classes the gtk-doc comments contained several errors, so it's
better to have no comments at all than erroneous comments.

The comment for _dh_window_display_uri() is restored to the state before
commit 6e745075da7a5713646fa9f02831c0775fbac500.