App: first step to make it private
DhApp is not re-usable. DhApp is a subclass of GtkApplication. An
application can have only one GtkApplication instance, so DhApp in the
current state doesn't make sense as a library.

This is the first step to make DhApp private, it removes the class from
the public docs.

The gtk-doc comments are removed, because it is not really useful, and
in other classes the gtk-doc comments contained several errors, so it's
better to have no comments at all than erroneous comments.