Generate enum types for the application
dh-enum-types.[ch] are for the library.
Generate dh-enum-types-app.[ch] for the enums that are part of only the
application. Currently the only source file of the app is dh-main.c, so
there are no enums, but some files will be moved from the lib to the
app, so it'll be useful (for DhWindow).