Compile DhApp, DhAssistant and DhWindow in the app, not the lib
To remove those classes from the public library API.

DhApp is not re-usable. DhApp is a subclass of GtkApplication. An
application can have only one GtkApplication instance, so DhApp in the
current state doesn't make sense as a library.

DhAssistant and DhWindow depend on DhApp, so they are not re-usable
either, it was not possible to use them in another app.

dh-app.h was not #included in devhelp.h.

The classes will maybe come back to the library in the future, in a
better shape.