DhLanguage: make it private
It is used only by dh-book-manager.c. DhBookManager doesn't expose
publicly DhLanguage, it is used internally to count the number of books
in each programming language.

The gtk-doc comments are removed, because it is not really useful, and
in other classes the gtk-doc comments contained several errors, so it's
better to have no comments at all than erroneous comments.