Add dh_free_resources() public function
There is already a dh_init() function, so it makes sense to add the
symmetric.

Put it in dh-init.c. "init" is not really a great filename since it's
both for initialization and de-initialization, but I have trouble to
find a better name. dh-common.c was not great either in my opinion.