parser: fix/improve error messages and do not translate them
Why not translating them?

- The error messages are not displayed in the GUI, only in the terminal
or the journal.

- When looking at the *.po files, XML attributes like "title", "name"
and "link" are sometimes translated but they must not be translated. It
would have been possible to add translator comments to say that those
must not be translated, but even with that I think some translators
would have translated the XML attributes anyway.

- With the English sentences, it is easy to search in the source code
where the error comes from. Also when someone reports a bug, etc.

- Devhelp is an application for developers, all the API references are
written in English, I think the person seeing an error message will be
able to understand it if it's in English…