parser: more logical order in parser_start_node_cb()
The same order as in the index files (generated with gtk-doc). Also the
same order as the function implementations.

In general there are more keywords than chapters, so the reverse order
was maybe to scrounge a tiny bit of performance. But here it's just
reading one more boolean value in a condition, that's nothing compared
to I/O.