parser: set out parameters only in dh_parser_read_file()
I think it's more natural to store the data in a normal way in the
struct.

parser->book_tree and parser->book_node were redundant, they contained
the same thing. So keep only book_node.

The out parameters are set only when the whole index file is
successfully parsed, which is a nicer behavior.

Also if one day we want to do asynchronous I/O, it'll be easier to
change dh_parser_read_file() into async/finish functions. The out
parameters are provided in the finish() function.