DhLink: modify get_type_as_string() -> type_to_string()
The function is not used by gnome-builder or Anjuta, so we can break the
API.

dh_link_type_to_string() takes a DhLinkType argument. API-wise, it's
slightly less convenient to use for the two call sites in Devhelp, but
the function is more powerful, it can be used independently from DhLink.