Window: improve function to find equivalent local URI
- Rename function, to make it clear that we try to find a local URI.
- Remove @window parameter, not used.
- Rename variable filename -> relative_url, the same name as in DhLink
constructors (because it can contain an anchor).

More importantly, compare book ID on the DhBook, to skip the whole book
at once instead of comparing the book ID for every DhLink of the book.
(the name book_id is clearer, but the function in DhBook is still called
dh_book_get_name() (vs title)).