Window: move New Tab button on the right side of the headerbar
Remember, commit f07f4bada6d0a68b2f2828e1f1b5a47518a7078b moved the New
Tab action from the hamburger menu to the headerbar. In the headerbar,
it was placed on the left side, after the back/forward buttons.

It's me who did this change, but most of the time I still try to go to
the hamburger menu to open a new tab… Maybe it's because it's difficult
to change my habits. But I think it's not only that: when opening a new
tab, the new tab appears on the right side of the window. The New Tab
button was placed on the opposite side, and to make things worse it was
above the side panel (if visible), not above the html pages.

So by placing the New Tab button on the right side of the headerbar (on
the left of the hamburger menu), the button is closer to what the action
does. It's more logical.

Ideally GtkNotebook (or in the future a GtkStackSwitcher) should have a
[+] button to open a new tab, like in Firefox currently, but I don't
think it's easily achievable with GtkNotebook right now.