DhLink: avoid some memory padding inside struct
This was a recent regression caused by the previous commit in DhLink.
But there was no comments to explain the order.

The guint is 4 bytes long on my machine (x86_64), there was 4 bytes of
padding for it. Plus 6 bytes of padding after the bit fields, for a
total of 64 bytes for the DhLink. Now there is only 2 bytes of padding,
for a total of 56 bytes (on x86_64).

With the gtk-doc documentation installed on my computer, the DhLink
struct is allocated more than 60k times, so optimizing it is worth it.