DhLink: use a union inside the struct to use less memory
At commit 326d80ff551cc29cde3380e6de217d27217184de, the DhLink struct
size was 56 bytes on x86_64. Now it is 40 bytes, because two pointers
have been removed.

With all the gtk-doc docs installed on my system, the memory used for
that struct is reduced from 3.8MB to 2.7MB (67899 DhLink's).

It is probably not transcendental, but it's anyway nice to fix that
FIXME. Also, the union has the benefit that it self-documents the fact
that some fields are available only for book links and vice versa.