KeywordModel: increase number of MAX_HITS
This was a low-hanging fruit.

Searching "GFile" didn't return an exact match (too many results).

100 was a bit too few. With 1000 the GtkSearchEntry is still responsive
(everything is synchronous in Devhelp, so some care must be taken to not
freeze the GUI during a too long time, if there was no MAX_HITS, maybe
it would cause a responsiveness problem in the GtkSearchEntry).

https://bugzilla.gnome.org/show_bug.cgi?id=784652