Don't use gdk_screen_get_width/height() (deprecated)
It's more complicated with the new API…

It would be possible to override the GtkWidget::realize vfunc to call
dh_util_window_settings_restore(), but ::realize can be called several
times since it can be unrealized, so it would require to store a boolean
instance variable to know that ::realize has already been called once.
So, calling gtk_widget_realize() is simpler.