App: do not set GtkApplication:register-session property
For example gnome-photos don't set this property. I don't think it is
needed. The less code the better.

Setting this property in dh_app_new() is done since the beginning, when
Devhelp has been ported to GtkApplication in
commit 53f9a8ce0c26e26f18187db1b0683b67dd0c24af.

If you revert this commit to add back register-session, add a comment to
explain why (the documentation of GtkApplication:register-session could
be improved too).