Window: remove ::open-link signal
DhWindow is no longer part of the library, so it's possible to simplify
the code.

Removing the signal permits to remove dh-enum-types-app.

Only DH_OPEN_LINK_NEW_TAB was used.

The signal implementation for DH_OPEN_LINK_NEW_WINDOW was broken, since
it just created a new window, it didn't open the link in the new window.