App: better function to get active DhWindow
peek_first_window() was not a good name, since by reading the function
prototype we didn't know that it created a new DhWindow if there was
none. And we didn't know that it returned a DhWindow.

get_active_main_window() has a name similar to
gtk_application_get_active_window(), and with the create_if_none
parameter things are done explicitly. create_if_none is set to FALSE in
several cases, for example to set the transient-for of secondary
windows.