Book: fix FIXME about page: and book: completions
See:
https://bugzilla.gnome.org/show_bug.cgi?id=791442#c3

The page: and book: filtering feature is not intended to be used with
the Devhelp application GUI.

In case of DH_LINK_TYPE_BOOK it would have been easy to fix the
completion, it's the book ID that needs to be completed, not the book
title, so it would have been possible to fix the code by calling
dh_link_get_book_id(). In case of DH_LINK_TYPE_PAGE there is no dh_link
function to get the page_id, so it would have required a bit more work
to implement dh_link_get_page_id().