parser: print warning if index file format version 1 still used
Now the format version 1 is officially deprecated. The plan is to remove
its support in the future, if it gets in the way. For example if one day
the version 3 is created, then the support for version 1 can definitely
be removed, supporting two versions in parallel is enough.

With the API docs installed on my system, only the libxml2 still uses
the format version 1.