BookManager: better memory management of timeout sources
Instead of keeping the DhBookManager alive (by reffing it) during the
duration of the timeout, do not ref the DhBookManager and store the
timeout source IDs to remove them in dispose().

It's currently not very useful, since DhBookManager is a singleton, so
it is anyway always alive. But I plan to factor out some classes from
DhBookManager, for example creating a DhBookListDirectory, taking care
of one books directory (so containing one GFileMonitor for that
directory). And such DhBookListDirectory objects could be created and
destroyed depending on the context (e.g. load books for a specific SDK
version, and have different directories for the different versions). So
with this commit it'll be easier to factor out the DhBookListDirectory
class, with the code already almost there with preparation work in
DhBookManager.