BookManager: remove dead code and remove DhLanguage
All of that is dead code, once the DhBookManager ::language-enabled and
::language-disabled have been removed (see previous commit).

That code contained a bug, see:
https://bugzilla.gnome.org/show_bug.cgi?id=791127
(a failed assertion), so now it is fixed.

DhLanguage was used only internally by DhBookManager. DhBookManager
didn't expose the priv->languages instance variable, it was used only to
send the signals, but the signals were not used.