Use DhCompletion
The GCompletion object was created in each DhSidebar (so it was
duplicated when there were several main windows). And DhSidebar needed
to maintain the GCompletion when books are added/removed.

Now the GList of DhCompletion objects is created on the fly, with each
DhCompletion object stored and created lazily in DhBook.

Devhelp now compiles without any deprecation warning.