Book: create DhCompletion when book enabled
A DhBook is initially enabled, but dh_book_set_enabled() is anyway
called after the DhBook is created and when it is enabled in the
preferences.

When GCompletion was used it was done similarly, the GCompletion was
created upfront, and at the time of the first completion the GCompletion
was already created.