KeywordModel: delegate DhLink matching to DhSearchContext
The implementation is different.

Now only two GPatternSpec's are created (prefix and nonprefix), not for
each keyword separately. All keywords are concatenated with the '*' glob
in-between.

For example searching:
"gtk_source_region region" is now equivalent to:
"gtk_source_region*region".

It's at least the behavior that I expect when I do a search with several
keywords.

https://bugzilla.gnome.org/show_bug.cgi?id=782257