KeywordModel: document "book:" and "page:" feature
Now I think that all the bugs with "book:" and "page:" are fixed (there
were 3 bugs related to that feature, it was completely broken). And with
the DhSearchContext helper class the code is hopefully easier to
understand (and on top of that DhSearchContext is unit-tested).

Document "book:" and page: in the API/developer doc of Devhelp, not the
user Mallard doc, because the feature is not intended to be used
directly by end users.

https://bugzilla.gnome.org/show_bug.cgi?id=791442