Window: simplify code to update window title
And connect to GtkNotebook::switch-page only once, when the change is
done, so that calling window_get_active_web_view() works.

Note that when the HTML page is blank, the title is now set to "Devhelp"
(before this commit, in that case there was no title at all, which was a
bit strange in my opinion).