Window: better window title when empty page
The doc of gtk_header_bar_set_title() doesn't recommend to have the app
name in the title.

"Empty Page" is already used for the tab title (when there are several
tabs), so I think it's a good idea to be consistent and have the window
title match the current tab title.