Copy TeplInfoBar from the Tepl library
It will be used internally by DhTab.

The code is copied from Tepl:
https://wiki.gnome.org/Projects/Tepl
(it's a library above GtkSourceView, so it's not great to depend on it
in Devhelp).

Modifications:
- in tepl-info-bar.h to remove
  the check "Only <tepl/tepl.h> can be included directly."
- add a message on top of both files to say that it's a copy (if it was
  not already obvious).
- in tepl-info-bar.c change the first /** to /* to avoid a warning when
  gtk-doc compiles the doc.

Not all the functions will be used, but I think it's better to keep all
the code, to make it easier to update it to a new version.