Window: delegate InfoBar handling to DhTab, and use TeplInfoBar
TeplInfoBar is so much nicer, it simplifies a lot the code.

Also, create each time a new TeplInfoBar widget, do not hide and re-use
the same. To ensure that the previous content is fully destroyed.

Destroy the info bar when a new page is loaded in the WebView, the code
was broken in DhWindow, there was the following bug: first error -> info
bar shown (OK) -> click on close button -> trigger another error -> info
bar *not* shown… Now it works fine.