WebView: add get_devhelp_title()
DhWebView is a subclass of WebKitWebView, so if later DhWebView is added
to the libdevhelp, the method name must not conflict with its parent
classes (for GObject Introspection support). There is already
webkit_web_view_get_title().

dh_web_view_get_devhelp_title() will also be used for the tab label. To
avoid code duplication.