Window: adapt GtkNotebook handling to make it work with DhTabLabel
When the close button is clicked in DhTabLabel, DhTabLabel destroys the
DhTab with gtk_widget_destroy(). Before, window_close_tab() was called,
which was responsible to update the GtkNotebook state accordingly. So
now the solution is to connect to more GtkNotebook signals, to show or
hide tabs, and to destroy the DhWindow when the last tab is closed.