Call gtk_window_close(), not gtk_widget_destroy(), to save GSettings
Saving the window GSettings is now done in the ::delete-event handler,
and apparently it is not called with gtk_widget_destroy().

This commit fixes the problem in those places:
- closing the last tab of a window with Ctrl+w.
- pressing Escape in the assistant.

There is also the problem for the app.quit GAction (Ctrl+q), but calling
gtk_window_close() there creates an infinite loop (with a lot of I/O
going on, I needed to reboot my computer by force :-( it doesn't happen
often to me that kind of thing, when programming).