App: save GSettings before g_application_quit()
When calling gtk_widget_destroy() on a GtkWindow the ::delete-event
signal is not emitted, so the GSettings were not saved when launching
the app.quit GAction (e.g. with Ctrl+q, or via the menu item).

Replace the loop of gtk_widget_destroy() by the simpler
g_application_quit() call. It's less friendly to memory debugging tools,
but at least that way we are sure that the app has quit. If
g_application_hold() is called somewhere in the code without the
corresponding g_application_release(), then the only solution to quit
the app was to kill the process.