Move library to its own directory, to separate it from the app
It's a first step: everything that is part of the library is moved to
the devhelp/ directory. The app stays in src/. With the devhelp/
directory name it'll be possible to have better #includes in the public
headers: #include <devhelp/dh-something.h>, instead of #include
"dh-something.h" (it'll be done by a later commit).

By better separating the lib from the app, we see that some stuff that
is currently built in the lib should actually be part of the app. Like
some *.ui files… It'll be fixed by later commits.

A next step will be to only #include <devhelp/devhelp.h> in the app when
accessing the lib, instead of #including individual headers.

https://bugzilla.gnome.org/show_bug.cgi?id=784351