meson: remove unused variable
In docs/reference/meson.build there is the ignore_headers parameter, but
it also contains the list of headers in devhelp/future/.

After this commit there is no code duplication. Another way would be to
create a global variable and populate it in devhelp/ and
devhelp/future/, to use it in docs/reference/. There is no perfect
solution, global variables should be avoided if possible, but it would
be better to list the *.h files in their corresponding directory.