AppData: update screenshots
I think the link to the wiki didn't work, the screenshot for the flatpak
is missing in gnome-software. The wiki is anyway not a great place to
host screenshots for the AppData. Hosting it in the git repo is not
great either because it makes the git repo bigger, it takes a longer
time to git clone, to run git gc, etc, and the resulting URL is not nice
since it requires the git service on the server (it's not just a static
file). So I've chosen people.gnome.org, if in the future I'm no longer
around the URL can anyway be changed.