Preferences: bind to the DhSettings:group-books-by-language property
It is planned to remove dh_settings_peek_contents_settings() from the
libdevhelp API, to not expose the GSettings objects.

Small regression: g_settings_bind() establishes a binding between the
writability of the GSettings key and the "sensitive" property of the
preferences widget. With g_object_bind_property() it is no longer the
case. I don't think that for Devhelp this is an important feature. If in
the future someone complains and the feature is really wanted, with a
valid use-case, then DhSettings can expose publicly the GSettings
objects and g_settings_bind() can again be used.