BookManager: remove book from list *before* sending ::book-deleted
It is more logical in that order. Ideally there should be a vfunc and it would
be the object method handler which removes the DhBook from the list (since it's
a G_SIGNAL_RUN_LAST signal). But I'm planning to refactor DhBookManager soon
with DhBookSelection, so it is not worth it.

Before commit f36eed21b8f14b09df23fbb4488b29cc63e781bb everything was
fine I think in the Devhelp application.

commit f36eed21b8f14b09df23fbb4488b29cc63e781bb changed DhPreferences
and calls dh_book_manager_get_books() inside its ::book-deleted handler.
So there was a bug. It is now fixed.