BookManager: remove all the implementation
There is now a new infrastructure: DhProfile, DhSettings and DhBookList.

The other classes in Devhelp have been ported to the new infrastructure.
Now the DhBookManager is empty and doesn't send signals. Some of the
DhBookManager API is still used in some IDEs, so to not break those apps
we must not remove completely the API.

But if IDEs are not adapted they will still work fine, because the
default DhProfile will be used in that case, which corresponds to how
DhBookManager was implemented (it'll be the same content).