Book: remove the "enabled" property
It was not really a property, but it's simpler to describe it this way.

Whether a DhBook is enabled is now decoupled from DhBook. For a book to
be enabled, it needs to be part of a DhBookList now. It's more flexible,
because there can be several different DhBookList's at the same time,
for example to create different profiles.

The DhCompletion objects are created lazily when calling
dh_book_get_completion(), to avoid creating them unnecessarily. But they
need to be created for the relevant DhBooks before the first completion
in the GtkSearchEntry. So do that in DhSidebar now (where the
GtkSearchEntry completion is implemented).