Window: delegate ::button-press-event handling to DhWebView
I've tried to fix the FIXME by handling ::button-press-event on the
whole GtkWindow, but it didn't work well, it was completely buggy, for
example it was necessary to do a double button click when the
GtkSearchEntry had the focus, a single button click didn't work. And
when the DhWebView had the focus the buttons didn't work at all.
Strange.