App: fix inconsistencies for zoom action descriptions
In the GtkShortcutsWindow it was "Zoom in" etc while in the menu it's
"Larger Text" etc. In gedit it's also "Larger Text" in the menu, but
gedit doesn't document those actions in its GtkShortcutsWindow.

So have "Larger text" etc consistently, in both the menu and the
GtkShortcutsWindow, to avoid confusion.