Settings: move font settings to the libdevhelp
It's the first step to be able to move DhWebView to the libdevhelp.

Loads of boilerplate :-( Handling GSettings in a library is not that
easy, I'm not entirely satisfied with DhSettings. A better solution
would maybe to have a code generation tool that reads a *.gschema.xml
file and creates GVariant properties, convenient getters/setters, bind
functions etc.