DconfMigration: improve API
Just rename free() to sync_and_free(), since the sync is important to
do (so it's better to be aware of it).

And move the function in such a manner that it is the same order as how
the functions should be called (the sync_and_free() is done at the end).

Note, DhDconfMigration is actually not used by Devhelp, it was written
as an experiment for future-proofness of DhSettings. But
DhDconfMigration *is* used by GNOME LaTeX, to handle the app rename
(LaTeXila -> GNOME LaTeX). But this place is the upstream place for
DhDconfMigration.