hacking: write 'Devhelp philosophy' section
Note that I'm far from the original author(s) of Devhelp, this is rather
my understanding after refactoring the whole codebase (for devhelp being
a generic tool; this was also the opinion of Frédéric Péters written in
a bugzilla comment somewhere, and over drinks in Brussels).

There was already a Devhelp library used by Anjuta, I've just improved a
bit the situation :-) And having a flexible library can be powerful, it
enables new potential, new low-hanging fruits. The code is cleaner, more
maintainble, it lowers the bar to entry for contributors because they
can read the API reference of the libdevhelp… in Devhelp.