build: do not try to mimic AX_COMPILER_FLAGS
It's easier to use the warning_level. It's not equivalent, but it's good
enough and it simplifies a lot the meson.build file, and is what is
recommended to do.

There was a Meson warning:
WARNING: Consider using the builtin warning_level option instead of
adding warning flags by hand.

I've tried warning_level=3 first, but it gives warnings of the following
types:
- [-Wpedantic] (some of them are easy to fix and it would be nice to fix
them).
- [-Woverlength-strings] in dh-resources.c.